# contracts/extensions/v1alpha1/provider-contracts.openapi.yaml
openapi: 3.1.0
info:
  title: Epydios AgentOps Control Plane Extension Provider Contracts
  version: v1alpha1
  summary: Versioned HTTP/JSON contracts for PolicyProvider, EvidenceProvider, and ProfileResolver.
  description: |
    This contract is the public plug-in surface used by OSS baseline providers and private
    providers (including AIMXS integrations). Breaking changes require a new version.
servers:
  - url: https://provider.internal
...
paths:
...
  /v1alpha1/evidence-provider/record:
    post:
      tags: [EvidenceProvider]
      operationId: evidenceRecord
      summary: Record a structured evidence event or artifact reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceRecordRequest"
      responses:
        "200":
          description: Record accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRecordResponse"
        "400":
          $ref: "#/components/responses/ProviderError"
        "500":
          $ref: "#/components/responses/ProviderError"
  /v1alpha1/evidence-provider/finalize-bundle:
    post:
      tags: [EvidenceProvider]
      operationId: evidenceFinalizeBundle
      summary: Finalize and materialize an evidence bundle manifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceFinalizeBundleRequest"
      responses:
        "200":
          description: Bundle finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceFinalizeBundleResponse"
        "400":
          $ref: "#/components/responses/ProviderError"
        "500":
          $ref: "#/components/responses/ProviderError"
...
components:
  responses:
    ProviderError:
      description: Structured provider error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProviderError"
  schemas:
    ObjectMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: Correlation ID generated by the control plane
        timestamp:
          type: string
          format: date-time
        tenantId:
          type: string
        projectId:
          type: string
        environment:
          type: string
        actor:
          type: object
          additionalProperties: true
...
    ProviderError:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
        message:
          type: string
        retryable:
          type: boolean
          default: false
        details:
          type: object
          additionalProperties: true
...
    EvidenceRecordRequest:
      type: object
      required: [meta, eventType]
      properties:
        meta:
          $ref: "#/components/schemas/ObjectMeta"
        eventType:
          type: string
        eventId:
          type: string
        runId:
          type: string
        stage:
          type: string
        payload:
          type: object
          additionalProperties: true
        artifactRefs:
          type: array
          items:
            type: object
            properties:
              uri:
                type: string
                format: uri
              checksum:
                type: string
              mediaType:
                type: string
        retentionClass:
          type: string
    EvidenceRecordResponse:
      type: object
      required: [evidenceId, accepted]
      properties:
        accepted:
          type: boolean
          default: true
        evidenceId:
          type: string
        checksum:
          type: string
        storageUri:
          type: string
        bundleHint:
          type: string
    EvidenceFinalizeBundleRequest:
      type: object
      required: [meta, bundleId]
      properties:
        meta:
          $ref: "#/components/schemas/ObjectMeta"
        bundleId:
          type: string
        runId:
          type: string
        evidenceIds:
          type: array
          items:
            type: string
        retentionClass:
          type: string
        annotations:
          type: object
          additionalProperties: true
    EvidenceFinalizeBundleResponse:
      type: object
      required: [bundleId, manifestUri]
      properties:
        bundleId:
          type: string
        manifestUri:
          type: string
        manifestChecksum:
          type: string
        itemCount:
          type: integer
          minimum: 0
...